backend_v2/requirements.txt
Flask==1.0.2
requests==2.21.0
google-cloud-firestore==0.31.0
